# src/main/java/com/marketing/web/swagger/swagger.yaml
swagger: '2.0'
info:
  description: OnlinePlasiyer api documentation
  version: '1.0'
  title: OnlinePlasiyer Documentation
  contact: {}
host: 'localhost:8080'
basePath: /
...
paths:
...
  '/api/categories/{id}':
    get:
      tags:
        - categories-controller
      summary: Find category by uuid
      operationId: findByUUIDUsingGET
      description: Every user can use
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ReadableCategory'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
...
  /api/users/activeStates:
    get:
      tags:
        - user-controller
      summary: Get user's active states
      operationId: getActiveStatesUsingGET
      description: Used only for user who has merchant role
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /api/users/addActiveState:
    post:
      tags:
        - user-controller
      summary: Add state in user's active states
      operationId: addActiveStateUsingPOST
      description: Used only for user who has merchant role
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: states
          description: states
          required: true
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: OK
          schema:
            type: string
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /api/users/customers:
    get:
      tags:
        - user-controller
      summary: get All Customers(Active and passive)
      operationId: getAllCustomersUsingGET
      description: Used only for user who has admin role
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/CustomerUser'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /api/users/customers/active:
    get:
      tags:
        - user-controller
      summary: get Active Customer Users
      operationId: getActiveCustomerUsersUsingGET
      description: Used only for user who has admin role
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/CustomerUser'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /api/users/customers/passive:
    get:
      tags:
        - user-controller
      summary: get Passive Customer Users
      operationId: getPassiveCustomerUsersUsingGET
      description: Used only for user who has admin role
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/CustomerUser'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /api/users/getMyInfos:
    post:
      tags:
        - user-controller
      summary: get logged in user's informations
      operationId: getUserInfosUsingPOST
      description: Every user can use
      consumes:
        - application/json
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/UserInfo'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /api/users/merchant:
    get:
      tags:
        - user-controller
      summary: get All Merchants(Active and passive)
      operationId: getAllMerchantsUsingGET
      description: Used only for user who has admin role
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/MerchantUser'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /api/users/merchant/active:
    get:
      tags:
        - user-controller
      summary: get ActiveMerchant Users
      operationId: getActiveMerchantUsersUsingGET
      description: Used only for user who has admin role
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/MerchantUser'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /api/users/merchant/passive:
    get:
      tags:
        - user-controller
      summary: get Passive Merchant Users
      operationId: getPassiveMerchantUsersUsingGET
      description: Used only for user who has admin role
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/MerchantUser'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  '/api/users/setActive/{id}':
    post:
      tags:
        - user-controller
      summary: Set passive user into active
      operationId: setActiveUserUsingPOST
      description: Used only for user who has admin role
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: string
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  '/api/users/setPassive/{id}':
    post:
      tags:
        - user-controller
      summary: set passive user into active
      operationId: setPassiveUserUsingPOST
      description: Used only for user who has admin role
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: string
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
...
  /sign-up:
    post:
      tags:
        - user-controller
      summary: signUp
      operationId: signUpUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: writableRegister
          description: writableRegister
          required: true
          schema:
            $ref: '#/definitions/WritableRegister'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ReadableRegister'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /signin:
    post:
      tags:
        - user-controller
      summary: login
      operationId: loginUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: writableLogin
          description: writableLogin
          required: true
          schema:
            $ref: '#/definitions/WritableLogin'
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
definitions:
  Address:
    type: object
    properties:
      city:
        type: string
      details:
        type: string
      id:
        type: integer
        format: int64
      state:
        type: string
      uuid:
        type: string
        format: uuid
    title: Address
...
  CustomerUser:
    type: object
    properties:
      address:
        $ref: '#/definitions/Address'
      email:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        type: boolean
      taxNumber:
        type: string
      username:
        type: string
    title: CustomerUser
  MerchantUser:
    type: object
    properties:
      activeStates:
        type: array
        items:
          type: string
      address:
        $ref: '#/definitions/Address'
      email:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        type: boolean
      taxNumber:
        type: string
      username:
        type: string
    title: MerchantUser
...
  ReadableCategory:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      parentId:
        type: string
      photoUrl:
        type: string
      subCategory:
        type: boolean
    title: ReadableCategory
...
  ReadableRegister:
    type: object
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        type: boolean
      taxNumber:
        type: string
      username:
        type: string
    title: ReadableRegister
...
  UserInfo:
    type: object
    properties:
      activeStates:
        type: array
        items:
          type: string
      address:
        $ref: '#/definitions/Address'
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
      username:
        type: string
    title: UserInfo
...
  WritableLogin:
    type: object
    properties:
      password:
        type: string
      username:
        type: string
    title: WritableLogin
...
  WritableRegister:
    type: object
    properties:
      city:
        type: string
      details:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      roleType:
        type: string
        enum:
          - MERCHANT
          - ADMIN
          - CUSTOMER
      state:
        type: string
      taxNumber:
        type: string
      username:
        type: string
    title: WritableRegister
...